+++ sites/partners/lib/helpers.ts
@@ -1,1 +1,52 @@
+import { t } from "@bloom-housing/ui-components"
+
+type DateTimePST = {
+  hour: string
+  minute: string
+  second: string
+  dayPeriod: string
+  year: string
+  day: string
+  month: string
+}
+
 export const emailRegex = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/
+
+export const convertDataToPst = (dateObj: Date) => {
+  if (!dateObj) {
+    return {
+      date: t("t.n/a"),
+      time: t("t.n/a"),
+    }
+  }
+
+  // convert date and time to PST
+  const ptFormat = new Intl.DateTimeFormat("en-US", {
+    timeZone: "America/Los_Angeles",
+    hour: "numeric",
+    minute: "numeric",
+    second: "numeric",
+    year: "numeric",
+    day: "numeric",
+    month: "numeric",
+  })
+
+  const originalDate = new Date(dateObj)
+  const ptDateParts = ptFormat.formatToParts(originalDate)
+  const timeValues = ptDateParts.reduce((acc, curr) => {
+    Object.assign(acc, {
+      [curr.type]: curr.value,
+    })
+    return acc
+  }, {} as DateTimePST)
+
+  const { month, day, year, hour, minute, second, dayPeriod } = timeValues
+
+  const date = `${month}/${day}/${year}`
+  const time = `${hour}:${minute}:${second} ${dayPeriod} PT`
+
+  return {
+    date,
+    time,
+  }
+}
